+++ crates/rubato-play/src/practice_configuration/ui.rs
@@ -139,7 +139,8 @@ impl PracticeConfiguration {
                     let times = &md.timeline_times;
                     if inc {
                         if !times.is_empty()
-                            && self.property.starttime + 2000 <= *times.last().expect("non-empty")
+                            && self.property.starttime as i64 + 2000
+                                <= *times.last().expect("non-empty")
                         {
                             self.property.starttime += 100;
                         }
@@ -157,7 +158,8 @@ impl PracticeConfiguration {
                     let times = &md.timeline_times;
                     if inc {
                         if !times.is_empty()
-                            && self.property.endtime <= *times.last().expect("non-empty") + 1000
+                            && (self.property.endtime as i64)
+                                <= *times.last().expect("non-empty") + 1000
                         {
                             self.property.endtime += 100;
                         }
